refactor: collapse lifecycle module into hl7v2
