+++ crates/hl7v2-lifecycle/Cargo.toml
@@ -12,11 +12,4 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-hl7v2-core = { workspace = true }
-hl7v2-model = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-chrono = { workspace = true, features = ["serde"] }
-sha2 = { workspace = true }
-
-[dev-dependencies]
-serde_json = { workspace = true }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = ["lifecycle"] }
