+++ crates/hl7v2/Cargo.toml
@@ -41,7 +41,7 @@ synthetic = [
     "dep:hl7v2-gen",
 ]
 redact = []
-lifecycle = ["dep:hl7v2-lifecycle"]
+lifecycle = ["dep:chrono", "dep:serde", "dep:sha2"]
 experimental-guard = ["dep:serde"]
 
 [dependencies]
@@ -53,6 +53,8 @@ hl7v2-parser = { version = "1.2.0", path = "../hl7v2-parser" }
 hl7v2-writer = { version = "1.2.0", path = "../hl7v2-writer" }
 thiserror = { workspace = true }
 serde = { workspace = true, features = ["derive"], optional = true }
+chrono = { workspace = true, features = ["serde"], optional = true }
+sha2 = { workspace = true, optional = true }
 
 hl7v2-json = { version = "1.2.0", path = "../hl7v2-json", optional = true }
 hl7v2-normalize = { version = "1.2.0", path = "../hl7v2-normalize", optional = true }
@@ -69,7 +71,9 @@ hl7v2-template-values = { version = "1.2.0", path = "../hl7v2-template-values", 
 hl7v2-faker = { version = "1.2.0", path = "../hl7v2-faker", optional = true }
 hl7v2-corpus = { version = "1.2.0", path = "../hl7v2-corpus", optional = true }
 hl7v2-gen = { version = "1.2.0", path = "../hl7v2-gen", optional = true }
-hl7v2-lifecycle = { version = "1.2.0", path = "../hl7v2-lifecycle", optional = true }
 
 [lints]
 workspace = true
+
+[dev-dependencies]
+serde_json = { workspace = true }
